adminpanel darkmode fix

+++ backend/controllers/items/itemCtrl.js
@@ -5,23 +5,25 @@ const cloudUploadImg = require("../utils/cloudConn")
 const fs = require("fs")
 
 const createItemCtrl = asyncHandler(async (req, res) => {
-  const { collectionId, description, title } = req.body
+  const { collectionId, description, title, userId } = req.body
 
   const localPath = `${req.file.filename}`
 
   const imgUploaded = await cloudUploadImg(localPath)
-  try {
-    const item = await Item.create({
-      title: title,
-      collectionId: collectionId,
-      description: description,
-      itemImg: imgUploaded?.url,
-    })
+  if (req.user.id === userId || req.user.isAdmin) {
+    try {
+      const item = await Item.create({
+        title: title,
+        collectionId: collectionId,
+        description: description,
+        itemImg: imgUploaded?.url,
+      })
 
-    fs.unlinkSync(localPath)
-    res.json(item)
-  } catch (error) {
-    res.json(error)
+      fs.unlinkSync(localPath)
+      res.json(item)
+    } catch (error) {
+      res.json(error)
+    }
   }
 })
 
@@ -70,15 +72,20 @@ const updateItemCtrl = asyncHandler(async (req, res) => {
 
 //delete item
 const deleteItemCtrl = asyncHandler(async (req, res) => {
-  const { id } = req.params
-  validateId(id)
-  try {
-    const item = await Item.findByIdAndDelete(id)
-    res.json(item)
-  } catch (error) {
-    res.json(error)
+  const paramsString = req.params.id.split(",")
+  const userId = paramsString[0]
+  const itemId = paramsString[1]
+
+  
+  if (req.user.id === userId || req.user.isAdmin) {
+    try {
+      const item = await Item.findByIdAndDelete(itemId)
+      res.json(item)
+    } catch (error) {
+      res.json(error)
+    }
+    res.json("Delete")
   }
-  res.json("Delete")
 })
 
 const toggleAddLikeToItemCtrl = asyncHandler(async (req, res) => {
@@ -88,10 +95,10 @@ const toggleAddLikeToItemCtrl = asyncHandler(async (req, res) => {
   //2. Find the login user
   const loginUserId = req?.user?._id
   //3. Find is this user has liked this item
-  console.log(loginUserId)
+
   const isLiked = item?.isLiked
   //4.Chech if this user has dislikes this item
-  
+
   //Toggle
   //Remove the user if he has liked the item
   if (isLiked) {
